Working version of the refactored controller
