+++ src/robot_control/nodes/controller
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
-from robot_control.controller import Controller
+from robot_control.controller_node import ControllerNode
 
 if __name__=="__main__":
-  Controller()
+  ControllerNode()
